+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## NEXT
 
+### Fixed
+* **core:** Datepicker bugs: Backspace zet cursor soms onterecht aan het einde van de input, MacOS Cmd + v en soortgelijke commands worden afgebroken ([#1597](https://github.com/dso-toolkit/dso-toolkit/issues/1597))
+
 ### Changed
 * **css** `dso-icons.svg`: Inline `<style>` elementen met `color: <kleur>` vervangen met `fill="<kleur>"` tbv. Content-security-policy issue met SVG spritesheet voor iconen ([#1600](https://github.com/dso-toolkit/dso-toolkit/issues/1600))
 * **dso-toolkit + css:** packages\dso-toolkit\src\styles\mixins\_grid.scss mergen met packages\sources\src\styling\grid\grid.scss ([#1644](https://github.com/dso-toolkit/dso-toolkit/issues/1644))